Created the BaseTag class which handles:     - checking for unknown attributes applied to a tag     - parsing the if/unless attributes for a tag
Updated all tags to use this base class: arg, env, group, include, launch, machine, node, param, remap, rosparam, and test.

Note: that roslaunch itself does not honor the if/unless attributes for the launch tag (although it parses them and does not produce a warning indicating that they are unknown attributes). The if/unless attributes the launch tag will be honored in this application.

Updated all code to honor the enabled/disabled state of tags throughout.