Initial support for launching remote processes. This uses the JSch library to execute commands over SSH.
Created the RosProcessIF class which provides a common interface between local and remote processes to make monitoring
 the two consistent.

Renamed the RosProcess class to RosLocalProcess.

Updated the FileLog class to be able to create custom log files rather than only using the singleton.

Added log files to all local processes. This includes passing the __log argument to the node itself, and managed updat
ing the log on respawns. This is for nodes configured using the 'log' output attribute.

Updated the StreamPrinter class to handle logging to a log file instead of just to the screen.