Added support for the - command line option which reads a launch file from stdin. This option will also work with the information printing options (--files, --nodes, --find-node, --args, --ros-args, etc). This is not the case with roslaunch but has been made to work with this application. If you use --find-node with a node that was specified in the file passed in stdin then it will print out a message indicating that rather than the filename (since the filename is not known).
Added documentation to the ArgumentParser class.

Reordered the handling of information printing messages so that no other messages are printed to the screen except just the desired data, e.g., if you pass the --files option it will only print the list of files and not any other messages.

Removed the warning that the --local option was not implemented, as an inital version of launching remote nodes is implemented.