Finished support for the - command line option.
Fixed a bug in the FileLog class that output was being buffered. Updated to flush the log when messages are written to it to ensure that all output is stored in the file.

Tracked down a bug with missing node output. The issue is caused by Java passing the output of programs to a file in order to retrieve the process output. As such, linux configures the processes to use a 4-kiB buffer for stdout which results in the stdout not being written to the file (and the Java roslaunch application never receiving the output). The current solution is to pass the node executables through 'stdbuf' which disables the buffering in favor of line buffering. This results in the stdout messages being received on the Java side much more reliably.

Refactored how node processes are being started to use the ProcessBuilder class which provides IO control that obsoletes the StreamPrinter class (which has now been deleted).